+++ RPG/printText.h
@@ -9,9 +9,9 @@ enum class Orders { ChangeWeapon };
 class printText
 {
 public:
-	printText(const sf::String& text, const sf::Font& font, Color textColor, unsigned int characterSize, pair<float, float> kordy);
-	printText(const sf::String& text, const sf::Font& font, Color textColor, unsigned int characterSize, pair<float, float> kordy, GameState state);
-	printText(const sf::String& text, const sf::Font& font, Color textColor, unsigned int characterSize, pair<float, float> kordy, Player* gracz, Orders metoda);
+	printText(const sf::String& text, const sf::Font& font, Color textColor, unsigned int characterSize, pair<int, int> kordy);
+	printText(const sf::String& text, const sf::Font& font, Color textColor, unsigned int characterSize, pair<int, int> kordy, GameState state);
+	printText(const sf::String& text, const sf::Font& font, Color textColor, unsigned int characterSize, pair<int, int> kordy, Player* gracz, Orders metoda);
 	Text& GetText();
 	GameState GetState() const;
 	Color& setColor();
